+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## Next
 
+* The meaning of `company-begin-commands` value t has slightly changed.
 * New transformer, `company-sort-by-backend-importance`.
 * When grouped back-ends are used, the back-end of the current candidate is
   indicated in the mode-line, enclosed in angle brackets.